[Fix] #141 - Add FloatPanel Package && Change BottomsheetView
