+++ CoPro/CoPro/Presentation/BottomTab/ViewController/BottomTabController.swift
@@ -11,7 +11,7 @@ import KeychainSwift
 
 class BottomTabController: UITabBarController {
     private func addTabBarSeparator() {
-        let separator = UIView(frame: CGRect(x: 0, y: 0, width: tabBar.frame.width, height: 1))
+        let separator = UITabBar(frame: CGRect(x: 0, y: 0, width: tabBar.frame.width, height: 1))
         separator.backgroundColor = UIColor.G3() // Set the color of the separator line
         tabBar.addSubview(separator)
     }
