+++ CoPro/CoPro/Presentation/Main/ViewController/MainViewController.swift
@@ -5,120 +5,35 @@
 //  Created by 문인호 on 12/27/23.
 //
 
+import FloatingPanel
 import UIKit
 
 import SnapKit
 import Then
 
-final class MainViewController: UIViewController {
+final class MainViewController: UIViewController, FloatingPanelControllerDelegate {
     
     //MARK: - UI Components
-    private let recruitVC = recruitViewController()
-    private let freeVC = freeViewController()
-    private let noticeVC = noticeViewController()
-
-    let grabberView = UIView()
-//    private lazy var noticeBoardView = UIView()
-    private let bottomSheetView: BottomSheetView = {
-        let view = BottomSheetView()
-        view.bottomSheetColor = UIColor.systemBackground
-        view.barViewColor = .G2()
-        view.layer.shadowColor = UIColor.black.cgColor
-        view.layer.shadowOffset = CGSize(width: 0, height: 1)
-        view.layer.shadowOpacity = 0.3
-        view.layer.shadowRadius = 10
-        return view
-    }()
-    var panGesture = UIPanGestureRecognizer()
+    var fpc = FloatingPanelController()
     private let scrollView = UIScrollView()
     private let pageControl = UIPageControl()
     var images: [UIImage] = [UIImage(named: "main_slider_image1")!, UIImage(named: "main_slider_image1")!, UIImage(named: "main_slider_image1")!] // 사용할 이미지들
-    private let segmentedControl: UISegmentedControl = {
-        let segmentedControl = UnderlineSegmentedControl(items: ["프로젝트", "자유", "공지사항"])
-      segmentedControl.translatesAutoresizingMaskIntoConstraints = false
-      return segmentedControl
-    }()
-    private lazy var pageViewController: UIPageViewController = {
-      let vc = UIPageViewController(transitionStyle: .scroll, navigationOrientation: .horizontal, options: nil)
-      vc.setViewControllers([self.dataViewControllers[0]], direction: .forward, animated: true)
-      vc.delegate = self
-      vc.dataSource = self
-      vc.view.translatesAutoresizingMaskIntoConstraints = false
-      return vc
-    }()
-    
-//    private lazy var addPostButton: UIButton = {
-//
-//        let btn = UIButton(frame: CGRect(x: 0, y: 0, width: 91, height: 37))
-//        btn.backgroundColor = UIColor.P2()
-//        btn.layer.cornerRadius = 20
-//        btn.setImage(UIImage(systemName: "plus"), for: .normal)
-//        btn.titleLabel?.font = .pretendard(size: 17, weight: .bold)
-//        btn.setTitle("글쓰기", for: .normal)
-//        btn.contentEdgeInsets = .init(top: 0, left: 1, bottom: 0, right: 1)
-//        btn.imageEdgeInsets = .init(top: 0, left: -1, bottom: 0, right: 1)
-//        btn.titleEdgeInsets = .init(top: 0, left: 1, bottom: 0, right: -1)
-//        btn.clipsToBounds = true
-//        btn.tintColor = .white
-//        btn.addTarget(self, action: #selector(addButtonDidTapped), for: .touchUpInside)
-//
-//        return btn
-//    }()
-      
-    var dataViewControllers: [UIViewController] {
-        [recruitVC, freeVC, noticeVC]
-    }
-    var currentPage: Int = 0 {
-        didSet {
-            // from segmentedControl -> pageViewController 업데이트
-            print(oldValue, self.currentPage)
-            let direction: UIPageViewController.NavigationDirection = oldValue <= self.currentPage ? .forward : .reverse
-            self.pageViewController.setViewControllers(
-                [dataViewControllers[self.currentPage]],
-                direction: direction,
-                animated: true,
-                completion: nil
-            )
-//            switch currentPage {
-//            case 0:
-//                // currentPage가 0이면 버튼을 보이게 함
-//                addPostButton.isHidden = false
-//                // a 페이지로 이동하는 코드
-//                break
-//            case 1:
-//                // currentPage가 1이면 버튼을 보이게 함
-//                addPostButton.isHidden = false
-//                // b 페이지로 이동하는 코드
-//                break
-//            case 2:
-//                // currentPage가 2이면 버튼을 숨김
-//                addPostButton.isHidden = true
-//                break
-//            default:
-//                break
-//            }
-        }
-    }
     
     //MARK: - LifeCycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
         setUI()
-        setLayout()
         setRegister()
-//        setGesture()
         setDelegate()
         setNavigate()
-        setAddTarget()
         setupScrollView()
         setupPageControl()
-        view.bringSubviewToFront(bottomSheetView)
-//        view.bringSubviewToFront(addPostButton)
+        setLayout()
     }
 }
 
-extension MainViewController: UIPageViewControllerDataSource, UIPageViewControllerDelegate, UIScrollViewDelegate {
+extension MainViewController: UIPageViewControllerDelegate, UIScrollViewDelegate {
     
     // MARK: - UI Components Property
 
@@ -127,29 +42,14 @@ extension MainViewController: UIPageViewControllerDataSource, UIPageViewControll
         scrollView.do {
             $0.showsHorizontalScrollIndicator = false
         }
-        segmentedControl.do {
-            $0.translatesAutoresizingMaskIntoConstraints = false
-            $0.selectedSegmentIndex = 0
-            $0.setTitleTextAttributes([NSAttributedString.Key.foregroundColor: UIColor.gray], for: .normal)
-            $0.setTitleTextAttributes(
-              [
-                NSAttributedString.Key.foregroundColor: UIColor.P2(),
-                .font: UIFont.pretendard(size: 17, weight: .bold)
-              ],
-              for: .selected
-            )
+        fpc.do {
+            $0.delegate = self
+            let contentVC = MainFloatingViewController()
+            $0.set(contentViewController: contentVC)
+//            $0.track(scrollView: contentVC.dataViewControllers[currentPage].tableView)
+            $0.addPanel(toParent: self)
+            $0.layout = MyFloatingPanelLayout()
         }
-        pageViewController.do {
-            $0.setViewControllers([self.dataViewControllers[0]], direction: .forward, animated: true)
-            $0.dataSource = self
-            $0.view.translatesAutoresizingMaskIntoConstraints = false
-        }
-        segmentedControl.do {
-            $0.setTitleTextAttributes([NSAttributedString.Key.foregroundColor: UIColor.G4(), .font: UIFont.pretendard(size: 17, weight: .bold)], for: .normal)
-            $0.setTitleTextAttributes([NSAttributedString.Key.foregroundColor: UIColor.P2(),.font: UIFont.pretendard(size: 17, weight: .bold)],for: .selected)
-            $0.selectedSegmentIndex = 0
-        }
-        self.changeValue(control: self.segmentedControl)
     }
     private func setNavigate() {
         let logoImage = UIImage(named: "logo_navigation")?.withRenderingMode(.alwaysOriginal)
@@ -160,45 +60,30 @@ extension MainViewController: UIPageViewControllerDataSource, UIPageViewControll
         self.navigationItem.leftBarButtonItem = leftButton
     }
     private func setLayout() {
-        view.addSubviews(bottomSheetView)
-        bottomSheetView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        bottomSheetView.addSubviews(segmentedControl, pageViewController.view)
+        self.view.addSubview(fpc.view)
+        // REQUIRED. It makes the floating panel view have the same size as the controller's view.
+        fpc.view.frame = self.view.bounds
+        fpc.view.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+          fpc.view.topAnchor.constraint(equalTo: self.view.topAnchor, constant: 0.0),
+          fpc.view.leftAnchor.constraint(equalTo: self.view.leftAnchor, constant: 0.0),
+          fpc.view.rightAnchor.constraint(equalTo: self.view.rightAnchor, constant: 0.0),
+          fpc.view.bottomAnchor.constraint(equalTo: self.view.bottomAnchor, constant: 0.0),
+        ])
+        // Add the floating panel controller to the controller hierarchy.
+        self.addChild(fpc)
 
-        segmentedControl.snp.makeConstraints {
-            $0.leading.equalTo(bottomSheetView.bottomSheetView.snp.leading)
-            $0.trailing.equalTo(bottomSheetView.bottomSheetView.snp.trailing)
-            $0.top.equalTo(bottomSheetView.bottomSheetView.snp.top).offset(10)
-            $0.height.equalTo(50)
+        // Show the floating panel at the initial position defined in your `FloatingPanelLayout` object.
+        fpc.show(animated: true) {
+            // Inform the floating panel controller that the transition to the controller hierarchy has completed.
+            self.fpc.didMove(toParent: self)
         }
-        pageViewController.view.snp.makeConstraints {
-            $0.leading.equalToSuperview().offset(4)
-            $0.trailing.equalToSuperview().offset(-4)
-            $0.bottom.equalToSuperview().offset(-4)
-            $0.top.equalTo(segmentedControl.snp.bottom).offset(5)
-        }
-//        view.addSubview(addPostButton)
-//        addPostButton.snp.makeConstraints {
-//            $0.trailing.equalToSuperview().offset(-16)
-//            $0.bottom.equalToSuperview().offset(-91)
-//            $0.width.equalTo(91)
-//            $0.height.equalTo(37)
-//        }
     }
     private func setDelegate() {
-        pageViewController.delegate = self
         scrollView.delegate = self
-        recruitVC.delegate = self
-        freeVC.delegate = self
-        noticeVC.delegate = self
     }
     private func setRegister() {
 
-    }
-    private func setAddTarget() {
-        self.segmentedControl.addTarget(self, action: #selector(changeValue(control:)), for: .valueChanged)
-        self.changeValue(control: self.segmentedControl)
     }
     func setupScrollView() {
         let screenWidth = UIScreen.main.bounds.width
@@ -267,49 +152,8 @@ extension MainViewController: UIPageViewControllerDataSource, UIPageViewControll
         pageControl.currentPage = Int(page)
     }
 
-    
-    func pageViewController(
-      _ pageViewController: UIPageViewController,
-      viewControllerBefore viewController: UIViewController
-    ) -> UIViewController? {
-      guard
-        let index = self.dataViewControllers.firstIndex(of: viewController),
-        index - 1 >= 0
-      else { return nil }
-      return self.dataViewControllers[index - 1]
-    }
-    func pageViewController(
-      _ pageViewController: UIPageViewController,
-      viewControllerAfter viewController: UIViewController
-    ) -> UIViewController? {
-      guard
-        let index = self.dataViewControllers.firstIndex(of: viewController),
-        index + 1 < self.dataViewControllers.count
-      else { return nil }
-      return self.dataViewControllers[index + 1]
-    }
-    func pageViewController(
-      _ pageViewController: UIPageViewController,
-      didFinishAnimating finished: Bool,
-      previousViewControllers: [UIViewController],
-      transitionCompleted completed: Bool
-    ) {
-      guard
-        let viewController = pageViewController.viewControllers?[0],
-        let index = self.dataViewControllers.firstIndex(of: viewController)
-      else { return }
-      self.currentPage = index
-      self.segmentedControl.selectedSegmentIndex = index
-    }
-
     // MARK: - @objc Method
 
-
-    @objc private func changeValue(control: UISegmentedControl) {
-      // 코드로 값을 변경하면 해당 메소드 호출 x
-      self.currentPage = control.selectedSegmentIndex
-    }
-    
     @objc func popToWriteViewController() {
         // 'Button 1'이 눌렸을 때의 동작을 여기에 작성합니다.
     }
@@ -318,25 +162,6 @@ extension MainViewController: UIPageViewControllerDataSource, UIPageViewControll
         let secondViewController = SearchBarViewController()
         self.navigationController?.pushViewController(secondViewController, animated: true)
     }
-    
-//    @objc func addButtonDidTapped() {
-//        switch currentPage {
-//            case 0:
-//            let addPostVC = AddProjectPostViewController()
-//                let navigationController = UINavigationController(rootViewController: addPostVC)
-//                navigationController.modalPresentationStyle = .fullScreen
-//                self.present(navigationController, animated: true, completion: nil)
-//                break
-//            case 1:
-//            let addPostVC = AddPostViewController()
-//                let navigationController = UINavigationController(rootViewController: addPostVC)
-//                navigationController.modalPresentationStyle = .fullScreen
-//                self.present(navigationController, animated: true, completion: nil)
-//                break
-//            default:
-//                break
-//            }
-//    }
 }
     
 extension MainViewController: RecruitVCDelegate {
@@ -346,4 +171,14 @@ extension MainViewController: RecruitVCDelegate {
             print("hello")
             navigationController?.pushViewController(detailVC, animated: true)
         }
+}
+
+class MyFloatingPanelLayout: FloatingPanelLayout {
+    let position: FloatingPanelPosition = .bottom
+    let initialState: FloatingPanelState = .half
+    let anchors: [FloatingPanelState: FloatingPanelLayoutAnchoring] = [
+        .full: FloatingPanelLayoutAnchor(absoluteInset: 16.0, edge: .top, referenceGuide: .safeArea),
+        .half: FloatingPanelLayoutAnchor(fractionalInset: 0.4, edge: .bottom, referenceGuide: .superview),
+        .tip: FloatingPanelLayoutAnchor(absoluteInset: 44.0, edge: .bottom, referenceGuide: .safeArea),
+    ]
 }
