+++ CoPro/CoPro/Presentation/Main/ViewController/noticeViewController.swift
@@ -131,9 +131,7 @@ extension noticeViewController: UITableViewDelegate, UITableViewDataSource {
             print(filteredPosts[indexPath.row].title)
             print("\(filteredPosts[indexPath.row].boardId)")
             detailVC.postId = filteredPosts[indexPath.row].boardId
-            let navigationController = UINavigationController(rootViewController: detailVC)
-            navigationController.modalPresentationStyle = .overFullScreen
-            self.present(navigationController, animated: true, completion: nil)
+            delegate?.didSelectItem(withId: filteredPosts[indexPath.row].boardId)
         } else {
             print("Invalid index")
             detailVC.postId = posts[indexPath.row].boardId
