+++ CoPro/CoPro/Presentation/Main/ViewController/recruitViewController.swift
@@ -170,10 +170,7 @@ extension recruitViewController: UITableViewDelegate, UITableViewDataSource {
             print(filteredPosts[indexPath.row].title)
             print("\(filteredPosts[indexPath.row].boardId)")
             detailVC.postId = filteredPosts[indexPath.row].boardId
-//            let navigationController = UINavigationController(rootViewController: detailVC)
-//            navigationController.modalPresentationStyle = .overFullScreen
-//            self.present(navigationController, animated: true, completion: nil)
-            self.navigationController?.pushViewController(detailVC, animated: true)
+            delegate?.didSelectItem(withId: filteredPosts[indexPath.row].boardId)
         } else {
             print("Invalid index")
             detailVC.postId = posts[indexPath.row].boardId
